Move ChatService to application layer
