+++ api/Avancira.Application/Extensions.cs
@@ -1,4 +1,5 @@
 ﻿using Avancira.Application.Catalog.Categories;
+using Avancira.Application.Catalog.Chats;
 using Avancira.Application.Services.Category;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -8,6 +9,7 @@ public static class Extensions
     public static IServiceCollection AddApplicationServices(this IServiceCollection services)
     {
         services.AddScoped<ICategoryService, CategoryService>();
+        services.AddScoped<IChatService, ChatService>();
 
         return services;
     }
