+++ api/Avancira.Infrastructure/Catalog/Extensions.cs
@@ -28,8 +28,6 @@ namespace Avancira.Infrastructure.Catalog
             services.AddTransient<ILessonCategoryService, LessonCategoryService>();
             services.AddTransient<IListingService, ListingService>();
             services.AddTransient<IUserService, UserService>();
-            services.AddTransient<IChatService, ChatService>();
-
             return services;
         }
     }
